Add benchmark Python script
The MODEL_NAME must still be set manually, as well as the API keys.

There is also no evaluation loop yet, because many models are not
sufficiently instruction following, necessitating better parsing. With a
single call, you can easily manually check whether the model's output is
correct and just not in the right format.